worked on recipe switch button logic
